+++ meta/core/dom/History.md
@@ -1,6 +1,22 @@
 
 # `dataMetaDom` Release history:
 
+## `1.0.6` - `2017-04-03 Mon` by [`mub`](https://github.com/mub)
+* Bug fix:
+    * Pojo Generator had autoversion parameter misspelled.
+    
+## `1.0.5` - `2017-03-06 Mon` by [`mub`](https://github.com/mub)
+* Updates:
+    * Added the URL and RAW types to MySQL generator
+    
+## `1.0.4` - `2017-02-28 Tue` by [`mub`](https://github.com/mub)
+* Updates:
+    * Adjusted Mapping and Bitset generation target package
+    
+## `1.0.3` - `2017-02-25 Sat` by [`mub`](https://github.com/mub)
+* Updates:
+    * Added new canned string validation pattern, `uuid`
+
 ## `1.0.2` - `2017-02-20 Mon` by [`mub`](https://github.com/mub)
 * Updates:
     * Fixed bug with the help STDOUT mixing up with STDERR output