+++ util/conv/protobuf/History.md
@@ -1,5 +1,9 @@
 # `dataMetaProtobuf` Release history:
 
+## `1.0.2` released `2017-04-04 Tue`
+* 1 update
+  * Upgraded to the core version `1.0.6`
+  
 ## `1.0.1` - `2017-02-20 Mon` by [`mub`](https://github.com/mub)
 * Update:
     * Bug fix: extraneous output to STDOUT